+++ prettierrc.js
@@ -1,6 +1,0 @@
-module.exports = {
-  semi: false,
-  trailingComma: "all",
-  printWidth: 120,
-  tabWidth: 2
-};