+++ src/__tests__/__fakes__/FakeComponentWithNoArgs.fake.ts
@@ -1,5 +1,5 @@
 export default class FakeComponentWithNoArgs {
   public get value() {
-    return 1
+    return 1;
   }
 }