+++ src/utils.ts
@@ -1,6 +1,7 @@
-import {IWorld} from './world/types'
-import World from './world/World'
+import { TBaseDependencies } from "types";
+import { IWorld } from "./world/types";
+import World from "./world/World";
 
-export const createWorld = <T extends object>(
-  dependencies: T
-): IWorld<T> => new World(dependencies)
+export const createWorld = <TDeps extends TBaseDependencies>(
+  dependencies: TDeps
+): IWorld<TDeps> => new World(dependencies);