+++ src/world/__tests__/World.test.ts
@@ -1,161 +1,187 @@
-import FakeComponentWithArgs from '../../__tests__/__fakes__/FakeComponentWithArgs.fake'
-import FakeComponentWithNoArgs from '../../__tests__/__fakes__/FakeComponentWithNoArgs.fake'
+import FakeComponentWithArgs from "../../__tests__/__fakes__/FakeComponentWithArgs.fake";
+import FakeComponentWithNoArgs from "../../__tests__/__fakes__/FakeComponentWithNoArgs.fake";
 import FakeSystem, {
-  TFakeSystemWithNoArgsComponents
-} from '../../__tests__/__fakes__/FakeSystem'
-import Entity from '../../entity/Entity'
-import {IEntity} from '../../entity/types'
-import {ISystem} from '../../system/types'
-import {IEcsWorld} from '../types'
-import World from '../World'
+  TFakeSystemComponents
+} from "../../__tests__/__fakes__/FakeSystem";
+import Entity from "../../entity/Entity";
+import { IEntity } from "../../entity/types";
+import { ISystem } from "../../system/types";
+import { IEcsWorld } from "../types";
+import World from "../World";
+import FakeSystemSingleComponent, {
+  TFakeSystemSingleComponents
+} from "../../__tests__/__fakes__/FakeSystemSingleComponent";
 
-describe('World', () => {
+type TTestDeps = {
+  test: string;
+};
+
+describe("World", () => {
   const deps = {
-    test: 'a'
-  }
+    test: "a"
+  };
 
-  type TDependencies = typeof deps
+  type TDependencies = typeof deps;
 
-  let world: IEcsWorld<TDependencies>
+  let world: IEcsWorld<TDependencies>;
 
   beforeEach(() => {
-    world = new World<TDependencies>(deps)
-  })
+    world = new World<TDependencies>(deps);
+  });
 
-  describe('addSystem', () => {
-    let system: ISystem<TFakeSystemWithNoArgsComponents>
-
-    beforeEach(() => {
-      world.addSystem(FakeSystem)
-      system = world.systems[0]
-    })
-
-    it('creates and adds system', () => expect(system).toBeDefined())
-  })
-
-  describe('createEntity', () => {
-    let entity: IEntity<TFakeSystemWithNoArgsComponents>
+  describe("addSystem", () => {
+    let system: ISystem<TFakeSystemComponents, TTestDeps>;
 
     beforeEach(() => {
-      entity = world.createEntity()
-    })
+      world.addSystem(FakeSystem);
+      system = world.systems[0];
+    });
 
-    it('creates valid entity', () => expect(entity).toBeInstanceOf(Entity))
+    it("creates and adds system", () => expect(system).toBeDefined());
+  });
 
-    it('getEntityComponents returns value for created entity', () => {
-      expect(world.getEntityComponents(entity)).toEqual(new Map())
-    })
-  })
-
-  describe('addEntityComponent', () => {
-    let entity: IEntity<TFakeSystemWithNoArgsComponents>
+  describe("createEntity", () => {
+    let entity: IEntity<TFakeSystemComponents>;
 
     beforeEach(() => {
-      entity = world.createEntity()
-    })
+      entity = world.createEntity();
+    });
 
-    describe('for component with initialization arguments', () => {
+    it("creates valid entity", () => expect(entity).toBeInstanceOf(Entity));
+
+    it("getEntityComponents returns value for created entity", () => {
+      expect(world.getEntityComponents(entity)).toEqual(new Map());
+    });
+  });
+
+  describe("addEntityComponent", () => {
+    let entity: IEntity<TFakeSystemComponents>;
+
+    beforeEach(() => {
+      entity = world.createEntity();
+    });
+
+    describe("for component with initialization arguments", () => {
       beforeEach(() => {
-        entity.addComponent(FakeComponentWithArgs, [0])
-      })
+        entity.addComponent(FakeComponentWithArgs, 0, "test");
+      });
 
-      it('getComponent in entity returns component', () => {
-        expect.assertions(1)
-        const component = entity.getComponent(FakeComponentWithArgs)
+      it("getComponent in entity returns component", () => {
+        expect.assertions(1);
+        const component = entity.getComponent(FakeComponentWithArgs);
 
         if (component) {
-          expect(component.args).toEqual(0)
+          expect(component.arg1).toEqual(0);
         }
-      })
-    })
+      });
+    });
 
-    describe('for component with no initialization arguments', () => {
+    describe("for component with no initialization arguments", () => {
       beforeEach(() => {
-        entity.addComponent(FakeComponentWithNoArgs)
-      })
+        entity.addComponent(FakeComponentWithNoArgs);
+      });
 
-      it('getComponent in entity returns component', () => {
-        expect.assertions(1)
-        const component = entity.getComponent(FakeComponentWithNoArgs)
+      it("getComponent in entity returns component", () => {
+        expect.assertions(1);
+        const component = entity.getComponent(FakeComponentWithNoArgs);
 
         if (component) {
-          expect(component.value).toBeDefined()
+          expect(component.value).toBeDefined();
         }
-      })
-    })
-  })
+      });
+    });
+  });
 
-
-  describe('removeEntityComponent', () => {
-    let entity: IEntity<TFakeSystemWithNoArgsComponents>
+  describe("removeEntityComponent", () => {
+    let entity: IEntity<TFakeSystemComponents>;
 
     beforeEach(() => {
-      entity = world.createEntity()
-    })
+      entity = world.createEntity();
+    });
 
-    describe('for component with initialization arguments', () => {
+    describe("for component with initialization arguments", () => {
       beforeEach(() => {
-        entity.addComponent(FakeComponentWithArgs, [0])
-        entity.removeComponent(FakeComponentWithArgs)
-      })
+        entity.addComponent(FakeComponentWithArgs, 0, "test");
+        entity.removeComponent(FakeComponentWithArgs);
+      });
 
-      it('getComponent in entity returns component', () => {
-        expect.assertions(1)
-        const component = entity.getComponent(FakeComponentWithArgs)
+      it("getComponent in entity returns component", () => {
+        expect.assertions(1);
+        const component = entity.getComponent(FakeComponentWithArgs);
 
-        expect(component).toBeUndefined()
-      })
-    })
-  })
+        expect(component).toBeUndefined();
+      });
+    });
+  });
 
-  describe('init', () => {
-    let systemInitSpy: jest.SpyInstance
-
-    beforeEach(async () => {
-      world.addSystem(FakeSystem)
-      systemInitSpy = jest.spyOn(world.systems[0], 'init')
-
-      await world.init()
-    })
-
-    it('initializes systems with dependencies', () =>
-      expect(systemInitSpy).toHaveBeenCalledWith(deps))
-  })
-
-  describe('update', () => {
-    let systemUpdateSpy: jest.SpyInstance
-    let systemPreUpdateSpy: jest.SpyInstance
-    let entity: IEntity<TFakeSystemWithNoArgsComponents>
+  describe("init", () => {
+    let systemInitSpy: jest.SpyInstance;
 
     beforeEach(async () => {
-      world.addSystem(FakeSystem)
-      entity = world
-        .createEntity<TFakeSystemWithNoArgsComponents>()
+      world.addSystem(FakeSystem);
+      systemInitSpy = jest.spyOn(world.systems[0], "init");
+
+      await world.init();
+    });
+
+    it("initializes systems with dependencies", () =>
+      expect(systemInitSpy).toHaveBeenCalledWith(deps));
+  });
+
+  describe("update", () => {
+    let system1UpdateSpy: jest.SpyInstance;
+    let system2UpdateSpy: jest.SpyInstance;
+    let system1PreUpdateSpy: jest.SpyInstance;
+    let entity1: IEntity<TFakeSystemComponents>;
+    let entity2: IEntity<TFakeSystemSingleComponents>;
+
+    beforeEach(async () => {
+      world.addSystem(FakeSystem);
+      world.addSystem(FakeSystemSingleComponent);
+      entity1 = world
+        .createEntity<TFakeSystemComponents>()
         .addComponent(FakeComponentWithNoArgs)
-        .addComponent(FakeComponentWithArgs)
-      systemUpdateSpy = jest.spyOn(world.systems[0], 'update')
-      systemPreUpdateSpy = jest.spyOn(world.systems[0], 'preUpdate')
-      await world.init()
-      world.update()
-    })
+        .addComponent(FakeComponentWithArgs, 0, "test");
+      entity2 = world
+        .createEntity<TFakeSystemSingleComponents>()
+        .addComponent(FakeComponentWithArgs, 0, "test");
+      system1UpdateSpy = jest.spyOn(world.systems[0], "update");
+      system1PreUpdateSpy = jest.spyOn(world.systems[0], "preUpdate");
 
-    it('runs preUpdate with deps', () =>
-      expect(systemPreUpdateSpy).toHaveBeenCalledWith(deps))
+      system2UpdateSpy = jest.spyOn(world.systems[1], "update");
+      await world.init();
+      world.update();
+    });
 
-    it('runs update with matching entity and deps', () =>
-      expect(systemUpdateSpy).toHaveBeenCalledWith(entity, deps))
-  })
+    it("runs preUpdate with deps", () =>
+      expect(system1PreUpdateSpy).toHaveBeenCalledWith(deps));
 
-  describe('removeEntity', () => {
-    let entity: IEntity<TFakeSystemWithNoArgsComponents>
+    it("tuns system 1 update only for matching entity", () =>
+      expect(system1UpdateSpy).toHaveBeenCalledTimes(1));
+
+    it("runs system 1 update with matching entity and deps", () =>
+      expect(system1UpdateSpy).toHaveBeenCalledWith(entity1, deps));
+
+    it("tuns system 2 update for both matching entities", () =>
+      expect(system2UpdateSpy).toHaveBeenCalledTimes(2));
+
+    it("runs system 2 update with matching entity and deps", () =>
+      expect(system2UpdateSpy).toHaveBeenNthCalledWith(1, entity1, deps));
+
+    it("runs system 2 update with matching entity and deps", () =>
+      expect(system2UpdateSpy).toHaveBeenNthCalledWith(2, entity2, deps));
+  });
+
+  describe("removeEntity", () => {
+    let entity: IEntity<TFakeSystemComponents>;
 
     beforeEach(() => {
-      entity = world.createEntity()
-      world.removeEntity(entity)
-    })
+      entity = world.createEntity();
+      world.removeEntity(entity);
+    });
 
-    it('getEntityComponents returns undefined for removed entity', () => {
-      expect(world.getEntityComponents(entity)).toEqual(undefined)
-    })
-  })
-})
+    it("getEntityComponents returns undefined for removed entity", () => {
+      expect(world.getEntityComponents(entity)).toEqual(undefined);
+    });
+  });
+});